moved instruction screen to new cpp
added header file
